Created an Email Testing Subdomain
Tested sending emails using flask mail. This service currently just has an input field for the recipient and the message, but it will be used to email a user their ticket information, as well as their "receipt".  Emails do send, but they require certain system variables to be set, as they require certain credentials.