Amusement, Game, and Ticketing Changes
The Amusements tab was properly filled out, the game has been updated via removing the "Restart" button and making the player press "Space" to start when they first load the page. The high score will also be displayed and stored in the player's browser. The ticketing part of the style sheet was updated as well. Also uploading a bunch of images to git, since they weren't uploaded to git before.